Replace `Exclude Name` with `Address Fields` setting
WooCommerce 9.8.0 includes the country by default when fetching the billing or shipping address. As this is an unexpected change, creators may not wish to include the country field in their subscriber’s custom fields. Rather than have another checkbox to exclude the country (similar to the exclude name checkbox), instead add an option to choose which parts of the address to include in the address custom fields, with default set to the data always expected.